+++ frontend/src/components/login/login.css
@@ -1,53 +1,78 @@
-.login {
-    color: #000;
-    width: 100%;
+html,
+.form-cont{
     height: 100%;
-    background-color: #fff;
-    flex-direction: column;
-    justify-content: space-between;
-    align-items: stretch;
-    font-family: Inter, Arial;
-    font-size: 1.5rem;
-    font-weight: 700;
-    line-height: 29px;
     display: flex;
-  }
-  
-  .cat-absolute-container {
-    z-index: 1;
-    width: 100%;
-    flex-direction: column;
-    flex-basis: 100%;
     justify-content: center;
     align-items: center;
-    margin-bottom: -738px;
-    padding-top: 46px;
-    display: flex;
+}
+
+.form-signin {
+  max-width: 330px;
+  padding: 15px;
+}
+
+.form-signin .form-floating:focus-within {
+  z-index: 2;
+}
+
+.form-signin input[type="email"] {
+  margin-bottom: -1px;
+  border-bottom-right-radius: 0;
+  border-bottom-left-radius: 0;
+}
+
+.form-signin input[type="password"] {
+  margin-bottom: 10px;
+  border-top-left-radius: 0;
+  border-top-right-radius: 0;
+}
+.bd-placeholder-img {
+  font-size: 1.125rem;
+  text-anchor: middle;
+  -webkit-user-select: none;
+  -moz-user-select: none;
+  user-select: none;
+}
+
+@media (min-width: 768px) {
+  .bd-placeholder-img-lg {
+    font-size: 3.5rem;
   }
-  
-  .enter-username {
-    margin-top: 23px;
-    margin-bottom: 64px;
-  }
-  
-  .enter-password {
-    margin-bottom: 88px;
-  }
-  
-  .log-in {
-    margin-bottom: 51px;
-    margin-left: 114px;
-  }
-  
-  .dont-have-an-account {
-    margin-top: 1px;
-    font-weight: 400;
-    text-decoration: underline;
-  }
-  
-  .login-and-sign-up-we {
-    width: 719px;
-    height: 738px;
-    object-fit: cover;
-  }
-  
+}
+
+.b-example-divider {
+  height: 3rem;
+  background-color: rgba(0, 0, 0, .1);
+  border: solid rgba(0, 0, 0, .15);
+  border-width: 1px 0;
+  box-shadow: inset 0 .5em 1.5em rgba(0, 0, 0, .1), inset 0 .125em .5em rgba(0, 0, 0, .15);
+}
+
+.b-example-vr {
+  flex-shrink: 0;
+  width: 1.5rem;
+  height: 100vh;
+}
+
+.bi {
+  vertical-align: -.125em;
+  fill: currentColor;
+}
+
+.nav-scroller {
+  position: relative;
+  z-index: 2;
+  height: 2.75rem;
+  overflow-y: hidden;
+}
+
+.nav-scroller .nav {
+  display: flex;
+  flex-wrap: nowrap;
+  padding-bottom: 1rem;
+  margin-top: -1px;
+  overflow-x: auto;
+  text-align: center;
+  white-space: nowrap;
+  -webkit-overflow-scrolling: touch;
+}
